Module aws_sdk_lookoutequipment::model

Modules

data_ingestion_job_summary

See DataIngestionJobSummary

data_pre_processing_configuration

See DataPreProcessingConfiguration

dataset_schema

See DatasetSchema

dataset_summary

See DatasetSummary

inference_execution_summary

See InferenceExecutionSummary

inference_input_configuration

See InferenceInputConfiguration

inference_input_name_configuration

See InferenceInputNameConfiguration

inference_output_configuration

See InferenceOutputConfiguration

inference_s3_input_configuration

See InferenceS3InputConfiguration

inference_s3_output_configuration

See InferenceS3OutputConfiguration

inference_scheduler_summary

See InferenceSchedulerSummary

ingestion_input_configuration

See IngestionInputConfiguration

ingestion_s3_input_configuration

See IngestionS3InputConfiguration

labels_input_configuration

See LabelsInputConfiguration

labels_s3_input_configuration

See LabelsS3InputConfiguration

model_summary

See ModelSummary

s3_object

See S3Object

tag

See Tag

Structs

DataIngestionJobSummary

Provides information about a specified data ingestion job, including dataset

DataPreProcessingConfiguration

The configuration is the TargetSamplingRate, which is the sampling rate of

DatasetSchema

Provides information about the data schema used with the given dataset.

DatasetSummary

Contains information about the specific data set, including name, ARN, and status.

InferenceExecutionSummary

Contains information about the specific inference execution, including input and output

InferenceInputConfiguration

Specifies configuration information for the input data for the inference, including S3

InferenceInputNameConfiguration

Specifies configuration information for the input data for the inference, including

InferenceOutputConfiguration

Specifies configuration information for the output results from for the inference,

InferenceS3InputConfiguration

Specifies configuration information for the input data for the inference, including

InferenceS3OutputConfiguration

Specifies configuration information for the output results from the inference,

InferenceSchedulerSummary

Contains information about the specific inference scheduler, including data delay

IngestionInputConfiguration

Specifies configuration information for the input data for the data ingestion job,

IngestionS3InputConfiguration

Specifies S3 configuration information for the input data for the data ingestion job.

LabelsInputConfiguration

Contains the configuration information for the S3 location being used to hold label

LabelsS3InputConfiguration

The location information (prefix and bucket name) for the s3 location being used for

ModelSummary

Provides information about the specified ML model, including dataset and model names and

S3Object

Contains information about an S3 bucket.

Tag

A tag is a key-value pair that can be added to a resource as metadata.

Enums

DataUploadFrequency

DatasetStatus

InferenceExecutionStatus

InferenceSchedulerStatus

IngestionJobStatus

ModelStatus

TargetSamplingRate