Module aws_sdk_lookoutequipment::operation

Structs

CreateDataset

Creates a container for a collection of data being ingested for analysis. The dataset

CreateInferenceScheduler

Creates a scheduled inference. Scheduling an inference is setting up a continuous

CreateModel

Creates an ML model for data inference.

DeleteDataset

Deletes a dataset and associated artifacts. The operation will check to see if any

DeleteInferenceScheduler

Deletes an inference scheduler that has been set up. Already processed output results

DeleteModel

Deletes an ML model currently available for Amazon Lookout for Equipment. This will prevent it from

DescribeDataIngestionJob

Provides information on a specific data ingestion job such as creation time, dataset

DescribeDataset

Provides a JSON description of the data that is in each time series dataset, including names, column names, and data types.

DescribeInferenceScheduler

Specifies information about the inference scheduler being used, including name, model,

DescribeModel

Provides a JSON containing the overall information about a specific ML model, including model name and ARN,

ListDataIngestionJobs

Provides a list of all data ingestion jobs, including dataset name and ARN, S3 location

ListDatasets

Lists all datasets currently available in your account, filtering on the dataset name.

ListInferenceExecutions

Lists all inference executions that have been performed by the specified inference

ListInferenceSchedulers

Retrieves a list of all inference schedulers currently available for your account.

ListModels

Generates a list of all models in the account, including model name and ARN, dataset,

ListTagsForResource

Lists all the tags for a specified resource, including key and value.

StartDataIngestionJob

Starts a data ingestion job. Amazon Lookout for Equipment returns the job status.

StartInferenceScheduler

Starts an inference scheduler.

StopInferenceScheduler

Stops an inference scheduler.

TagResource

Associates a given tag to a resource in your account. A tag is a key-value pair which

UntagResource

Removes a specific tag from a given resource. The tag is specified by its key.

UpdateInferenceScheduler

Updates an inference scheduler.