Struct aws_sdk_lookoutequipment::output::describe_model_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DescribeModelOutput

Implementations

impl Builder

pub fn model_name(self, input: impl Into<String>) -> Self

The name of the ML model being described.

pub fn set_model_name(self, input: Option<String>) -> Self

pub fn model_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the ML model being described.

pub fn set_model_arn(self, input: Option<String>) -> Self

pub fn dataset_name(self, input: impl Into<String>) -> Self

The name of the dataset being used by the ML being described.

pub fn set_dataset_name(self, input: Option<String>) -> Self

pub fn dataset_arn(self, input: impl Into<String>) -> Self

The Amazon Resouce Name (ARN) of the dataset used to create the ML model being described.

pub fn set_dataset_arn(self, input: Option<String>) -> Self

pub fn schema(self, input: impl Into<String>) -> Self

A JSON description of the data that is in each time series dataset, including names, column names, and data types.

pub fn set_schema(self, input: Option<String>) -> Self

pub fn labels_input_configuration(self, input: LabelsInputConfiguration) -> Self

Specifies configuration information about the labels input, including its S3 location.

pub fn set_labels_input_configuration(
self,
input: Option<LabelsInputConfiguration>
) -> Self

pub fn training_data_start_time(self, input: Instant) -> Self

Indicates the time reference in the dataset that was used to begin the subset of training data for the ML model.

pub fn set_training_data_start_time(self, input: Option<Instant>) -> Self

pub fn training_data_end_time(self, input: Instant) -> Self

Indicates the time reference in the dataset that was used to end the subset of training data for the ML model.

pub fn set_training_data_end_time(self, input: Option<Instant>) -> Self

pub fn evaluation_data_start_time(self, input: Instant) -> Self

Indicates the time reference in the dataset that was used to begin the subset of evaluation data for the ML model.

pub fn set_evaluation_data_start_time(self, input: Option<Instant>) -> Self

pub fn evaluation_data_end_time(self, input: Instant) -> Self

Indicates the time reference in the dataset that was used to end the subset of evaluation data for the ML model.

pub fn set_evaluation_data_end_time(self, input: Option<Instant>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of a role with permission to access the data source for the ML model being described.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn data_pre_processing_configuration(
self,
input: DataPreProcessingConfiguration
) -> Self

The configuration is the TargetSamplingRate, which is the sampling rate of the data after post processing by Amazon Lookout for Equipment. For example, if you provide data that has been collected at a 1 second level and you want the system to resample the data at a 1 minute rate before training, the TargetSamplingRate is 1 minute.

When providing a value for the TargetSamplingRate, you must attach the prefix "PT" to the rate you want. The value for a 1 second rate is therefore PT1S, the value for a 15 minute rate is PT15M, and the value for a 1 hour rate is PT1H

pub fn set_data_pre_processing_configuration(
self,
input: Option<DataPreProcessingConfiguration>
) -> Self

pub fn status(self, input: ModelStatus) -> Self

Specifies the current status of the model being described. Status describes the status of the most recent action of the model.

pub fn set_status(self, input: Option<ModelStatus>) -> Self

pub fn training_execution_start_time(self, input: Instant) -> Self

Indicates the time at which the training of the ML model began.

pub fn set_training_execution_start_time(self, input: Option<Instant>) -> Self

pub fn training_execution_end_time(self, input: Instant) -> Self

Indicates the time at which the training of the ML model was completed.

pub fn set_training_execution_end_time(self, input: Option<Instant>) -> Self

pub fn failed_reason(self, input: impl Into<String>) -> Self

If the training of the ML model failed, this indicates the reason for that failure.

pub fn set_failed_reason(self, input: Option<String>) -> Self

pub fn model_metrics(self, input: impl Into<String>) -> Self

The Model Metrics show an aggregated summary of the model's performance within the evaluation time range. This is the JSON content of the metrics created when evaluating the model.

pub fn set_model_metrics(self, input: Option<String>) -> Self

pub fn last_updated_time(self, input: Instant) -> Self

Indicates the last time the ML model was updated. The type of update is not specified.

pub fn set_last_updated_time(self, input: Option<Instant>) -> Self

pub fn created_at(self, input: Instant) -> Self

Indicates the time and date at which the ML model was created.

pub fn set_created_at(self, input: Option<Instant>) -> Self

pub fn server_side_kms_key_id(self, input: impl Into<String>) -> Self

Provides the identifier of the KMS key used to encrypt model data by Amazon Lookout for Equipment.

pub fn set_server_side_kms_key_id(self, input: Option<String>) -> Self

pub fn off_condition(self, input: impl Into<String>) -> Self

Indicates that the asset associated with this sensor has been shut off. As long as this condition is met, Lookout for Equipment will not use data from this asset for training, evaluation, or inference.

pub fn set_off_condition(self, input: Option<String>) -> Self

pub fn build(self) -> DescribeModelOutput

Consumes the builder and constructs a DescribeModelOutput